+++ lws.config.js
@@ -1,0 +1,12 @@
+'use strict';
+
+module.exports = {
+    rewrite: [
+        {
+            from: '/protocol/*',
+            to: '/docs/$1',
+        },
+    ],
+    directory: '.',
+    spa: '/docs/index.html',
+};