+++ site/routes/domain.vue
@@ -3,7 +3,7 @@
         <div class="domain__intro">
             <h1>{{ domain.id }}</h1>
             <div class="domain__description"
-                 v-html="domain.description">
+                 v-html="domain.spec.description">
             </div>
         </div>